Added custom model instantiater, and filtering.
Custom Model Instantiater allows the "consumer" to specify a function
which creates new model instances. This way the consumer can define the
schema of the objects used behind the lists.

Filtering allows the "consumer" to pass in a function(item) which returns
a boolean. This can be used to dermine whether objects from the server
should be shown in the list. This means that multiple lists can share the
same back-end url, but the items can be divided amongst the lists, based
on some attribute of the model (such as age for people).